Refactor: split god file into 5 focused modules + add README
server.ts (69 lines) — entry point, MCP setup, shutdown
src/config.ts (50 lines) — env loading, validation, exports
src/format.ts (63 lines) — event formatting + truncation
src/mute.ts (66 lines) — mute state, counters, per-repo/global/timed
src/verify.ts (18 lines) — HMAC-SHA256 signature verification
src/webhook.ts (143 lines) — HTTP server, control endpoints, webhook handler

README.md — full setup guide, security model, control endpoints

No functional changes. 17 tests, 43 assertions, all passing.
